functions/add_recommendations/requirements.txt
google-cloud-bigquery==1.28.0
pandas-gbq==0.14.0
pandas==1.1.3
requests==2.24.0
